+++ UtilityFace/UtilityFace/Props/Int64IdExtensions.cs
@@ -48,7 +48,7 @@ public static class Int64IdExtensions
     public static string Format(this Int64Id prop, params object[] values)
     {
         //Prefer friendly name if available?
-        if (prop.TryGetFriendly((int)values[0], out var friendly))
+        if (prop.TryGetFriendly((long)values[0], out var friendly))
             values[0] = friendly;
 
         //Use a format string if it exists?
